[add] enforce stricter support for implicit type conversion per C99
- Implements stricter integer/boolean conversion rules.
- `bool` values are now coerced to 0 or 1 on assignment (C99 §6.3.1.2).

+++ include/tokenizer.h
@@ -54,7 +54,7 @@ Token *peek_ident();
 Token *consume_ident();
 Token *expect_ident();
 Token *consume_string();
-bool is_number(long *result);
-long expect_number();
+bool is_number(long long *result);
+long long expect_number();
 
 #endif  // TOKENIZER_C_COMPILER
